requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.22
psycopg2-binary==2.9.9
WTForms==3.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
Werkzeug==2.2.3
email_validator==2.0.0 